generative_ai_tutorial/12.langsmith/requirements.txt
langchain
langchain-openai
langchain-chroma
python-dotenv
pypdf
fastapi
uvicorn
langchain_community